+++ docs/change-log.md
@@ -12,6 +12,7 @@ outline: deep
 * NOTE: Any dropzones you have in use need to be updated to not use a dot in the array of file extensions.
 * Implemented: figure out a way to share file extensions for s3files from schema #104
 * NOTE: Because of the above you may want to check out the new acceptedFileExtensions attribute in ving schemas and migrate your S3File integrations to use it.
+* Implemented: add a display of an s3file thumbnail to the page generator #105
 
 ## 2024-04-23
 * Created SelectInput component to replace FormSelect. However, you should use FormInput with type select instead of using this directly in most cases.